+++ lib/src/models/scan_filter.dart
@@ -21,7 +21,7 @@ class ManufacturerDataFilter {
   int companyIdentifier;
 
   // Mask and data must be of same length
-  Uint8List? data;
+  Uint8List? payload;
 
   /// For any bit in the mask, set it the 1 if it needs to match
   /// the one in manufacturer data, otherwise set it to 0.
@@ -30,12 +30,12 @@ class ManufacturerDataFilter {
 
   ManufacturerDataFilter({
     required this.companyIdentifier,
-    this.data,
+    this.payload,
     this.mask,
   });
 
   @override
   String toString() {
-    return 'ManufacturerDataFilter(companyIdentifier: $companyIdentifier, data: $data, mask: $mask)';
+    return 'ManufacturerDataFilter(companyIdentifier: $companyIdentifier, payload: $payload, mask: $mask)';
   }
 }